[proliferate] split workspace header tabs view model
