+++ desktop/src/components/workspace/shell/tabs/tab-rendering.tsx
@@ -8,7 +8,7 @@ import {
 import type {
   HeaderChatMenuEntry,
   HeaderChatTabEntry,
-} from "@/hooks/workspaces/tabs/use-workspace-header-tabs-view-model";
+} from "@/hooks/workspaces/tabs/workspace-header-tabs-view-model-types";
 
 export function renderChatTabIcon(
   tab: Pick<HeaderChatTabEntry | HeaderChatMenuEntry, "agentKind" | "viewState">,
